Processing GUI Simpler: audio on Alpha Detection
For the Processing_GUI_Simpler, I expanded the "Alpha Detection"
behavior.  Now, when Alpha Detection is selected, and if Alpha waves are
actually detected, a tone is played so that you know you're making
Alpha, even when your eyes are closed.  The pitch of the tone is
proportional to the SNR (ie, strength) of the alpha waves.   We might
want to make it proportional to just the amplitude and not the SNR, but
SNR seemed best at the moment.