Processing_GUI: improve binary receiving
Big-time rework of the routines to receive the binary data from OpenBCI
via the Arduino.  Removed the ability to do text transfer.  Now, only
does binary.  Tested for 8 channels.  Untested for 16.  Tested at 115200
kbps and at 2*115200.  Setting it to 115200 is more stable.  Changes
need to be ported to Processing_GUI_Simpler.