ADS1299Manager: improve init, bias, LOFF_FLIP
Improve initialization sequence and added initialization of of the
LEAD_OFF parameters.  Added LEAD_OFF FLIP to swap the polarity of the
injected LEAD_OFF signal based on V1 vs V2.  Finally, for V2, we now
generate the bias based on both the N and P inputs, per Joel's
suggestion.