ProcGUI: playback: fixed fast play.  now wraps around
Fixed bug where, if you hit stop, waited, then start, the speed of
playback would be fast until it caught up with the amount of time that
past while it was stopped.  I fixed this.  Also, when it hits the end of
the playback file, it now automatically starts back at the beginning
again....it now will loop continuously.